refactor(web): [Task in Team] Add search feature
